Moved to double-precision for sample ratios
When resampling long tracks, it’s a good idea to have at least 32 bits
of precision in the mantissa of the sample ratio/inverse index
multiplier so as to avoid time-domain aliasing problems to due with
limited precision. Ideally i would be using a fixed point type here,
but double is currently the best option.

+++ include/Red/Audio/AudioBuffer.h
@@ -33,7 +33,7 @@ namespace Red
 			AudioBuffer ( void * Data, AudioBufferType Type, uint32_t Channels, uint64_t SampleCount, Util :: IFunction1 <void, void *> * OnFree = NULL, Util :: IFunction1 <bool, AudioBuffer *> * OnUnReferenced = NULL );
 			
 			static AudioBuffer * CopyReformated ( AudioBuffer & Source, AudioBufferType NewDataType, uint32_t NewChannelCount = 0 );
-			static AudioBuffer * CopyReformatedResampled ( AudioBuffer & Source, ResampleMode Mode, AudioBufferType NewDataType, float SampleRatio, uint32_t NewChannelCount = 0 );
+			static AudioBuffer * CopyReformatedResampled ( AudioBuffer & Source, ResampleMode Mode, AudioBufferType NewDataType, double SampleRatio, uint32_t NewChannelCount = 0 );
 			
 			void * GetRawBuffer ();
 			
@@ -43,13 +43,13 @@ namespace Red
 			uint64_t GetSampleCount ();
 			
 			void BlitBuffer ( AudioBuffer & Source, uint32_t SourceChannel, uint64_t SampleCount, uint64_t SourceStartSample = 0, uint64_t TargetStartSample = 0, uint32_t TargetChannel = 0xFFFFFFFF );
-			void BlitBufferResampled ( AudioBuffer & Source, ResampleMode Mode, uint32_t Channel, uint64_t SampleCount, uint64_t SourceStartSample = 0, uint64_t TargetStartSample = 0, float SampleRatio = 1.0f, uint32_t TargetChannel = 0xFFFFFFFF );
+			void BlitBufferResampled ( AudioBuffer & Source, ResampleMode Mode, uint32_t Channel, uint64_t SampleCount, uint64_t SourceStartSample = 0, uint64_t TargetStartSample = 0, double SampleRatio = 1.0f, uint32_t TargetChannel = 0xFFFFFFFF );
 			
 			void AddBuffer ( AudioBuffer & Source, uint32_t SourceChannel, uint64_t SampleCount, uint64_t SourceStartSample = 0, uint64_t TargetStartSample = 0, uint32_t TargetChannel = 0xFFFFFFFF );
-			void AddBufferResampled ( AudioBuffer & Source, ResampleMode Mode, uint32_t Channel, uint64_t SampleCount, uint64_t SourceStartSample = 0, uint64_t TargetStartSample = 0, uint32_t TargetChannel = 0xFFFFFFFF, float SampleRatio = 1.0f );
+			void AddBufferResampled ( AudioBuffer & Source, ResampleMode Mode, uint32_t Channel, uint64_t SampleCount, uint64_t SourceStartSample = 0, uint64_t TargetStartSample = 0, uint32_t TargetChannel = 0xFFFFFFFF, double SampleRatio = 1.0f );
 			
 			void AddBufferScaled ( AudioBuffer & Source, float Scale, uint32_t SourceChannel, uint64_t SampleCount, uint64_t SourceStartSample = 0, uint64_t TargetStartSample = 0, uint32_t TargetChannel = 0xFFFFFFFF );
-			void AddBufferResampledScaled ( AudioBuffer & Source, float Scale, ResampleMode Mode, uint32_t Channel, uint64_t SampleCount, uint64_t SourceStartSample = 0, uint64_t TargetStartSample = 0, uint32_t TargetChannel = 0xFFFFFFFF, float SampleRatio = 1.0f );
+			void AddBufferResampledScaled ( AudioBuffer & Source, float Scale, ResampleMode Mode, uint32_t Channel, uint64_t SampleCount, uint64_t SourceStartSample = 0, uint64_t TargetStartSample = 0, uint32_t TargetChannel = 0xFFFFFFFF, double SampleRatio = 1.0f );
 			
 			void ClearBufferFloat ( uint32_t Channel, float Value = 0.0f, uint64_t StartSample = 0, uint64_t SampleCount = 0xFFFFFFFFFFFFFFFF );
 			void ClearBufferInt ( uint32_t Channel, int64_t Value = 0, uint64_t StartSample = 0, uint64_t SampleCount = 0xFFFFFFFFFFFFFFFF );
